refactor: UserHeaderNotFoundException 적용
header 값이 필요한 요청에 header가 null값이라면 예외처리

+++ src/main/java/com/user/management/controller/UserController.java
@@ -3,6 +3,7 @@ package com.user.management.controller;
 import com.user.management.dto.UserCreateRequest;
 import com.user.management.dto.UserLoginRequest;
 import com.user.management.entity.User;
+import com.user.management.exception.UserHeaderNotFoundException;
 import com.user.management.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -19,14 +20,20 @@ public class UserController {
     private final UserService userService;
 
     @GetMapping
-    public ResponseEntity<List<User>> findAllUsers(@RequestHeader("X-USER-ID") String id)
+    public ResponseEntity<List<User>> findAllUsers(@RequestHeader(value = "X-USER-ID", required = false) String id)
     {
+        if(id == null)
+            throw new UserHeaderNotFoundException();
+
         return ResponseEntity.ok().body(userService.getAllUsers(id));
     }
 
     @GetMapping("/myPage")
-    public ResponseEntity<User> findUser(@RequestHeader("X-USER-ID") String id)
+    public ResponseEntity<User> findUser(@RequestHeader(value = "X-USER-ID", required = false) String id)
     {
+        if(id == null)
+            throw new UserHeaderNotFoundException();
+
         return ResponseEntity.ok().body(userService.getUserById(id));
     }
 
@@ -44,15 +51,21 @@ public class UserController {
     }
 
     @PutMapping
-    public ResponseEntity<Void> updateUser(@RequestBody UserCreateRequest userCreateRequest, @RequestHeader("X-USER-ID") String id)
+    public ResponseEntity<Void> updateUser(@RequestBody UserCreateRequest userCreateRequest, @RequestHeader(value = "X-USER-ID", required = false) String id)
     {
+        if(id == null)
+            throw new UserHeaderNotFoundException();
+
         userService.updateUser(userCreateRequest, id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
     @DeleteMapping
-    public ResponseEntity<Void> deleteUser(@RequestHeader("X-USER-ID") String id)
+    public ResponseEntity<Void> deleteUser(@RequestHeader(value = "X-USER-ID", required = false) String id)
     {
+        if(id == null)
+            throw new UserHeaderNotFoundException();
+
         userService.deleteUser(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
